Adding IO#sync and IO#sync= while I'm in here
and making the stdio streams behave the same as in MRI.  Unfortunately
this results in having to explicitly set `sync = true` in some tests
until I can figure out (or have someone tell me) how to properly flush
when leaving a scope.  Not happy.

Oh, and finally removed the IO#sysopen exclusion in the mspec config.