Added test for stuff rubyspec doesn't cover
In the Math module of ruby 1.9.3 arguments such as '1' or '2.0' won't be
accepted as input arguments and will result in a TypeError if they do get
passed. However there is no rubyspec for this kind of behaviour (at least not
in rubyspec/core/math/) so I added a few unittests.