Added const ADDRESS_SIZE to Platform
In order to do this, Platform first has to be defined in
tests/modules/test_ffi.py.
This means that every time it is mentioned in lib-ruby/ffi it is reopened
instead of created. However, I didn't experience any problems with it. Seems
that monkey patching is already supported in topaz. However this is something
to keep in mind.