Added unit test for a bug I ran over
Inside a module the method defined? didn't work correctly.
The first part makes sure that when a variable @foo inside module A was not
defined at all defined?(@foo) will return false inside A.
The second part makes sure that when a variable @foo was defined as nil inside
A defined(@foo) will return true (and not false because @foo is nil).