Implemented W_FunctionTypeObject.invoke
It can invoke a proc with ll_args. The result is a ruby object.
The solution obviously copies lots of code from topaz.modules.ffi.function so
the next task is to somehow move the copied code into a separate module witch
both W_FunctionTypeObject and W_FunctionObject can use.