Moved code for enum from _unwrap_and_write_to_address to function
Handling the enum there is better because much more information is available
inside of a Function method. This will make refactoring the code easier and
maybe also make it possible to generalize the code more (which probably needs
to be done in the future because an Enum is not the only Mapped type).