Simplified code for handling Enum objects
Due to moving it to the function module this code could be simplified a lot.
The trick is that in function the Type object is still available, not only the
typeindex. This Type is a wrapper around the Enum class which means that it can
delegate lots of stuff to Enum.